todo_backend/requirements.txt
django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
